Module ic_crypto_internal_multi_sig_bls12381::types

BLS12-381 multisignature types.

Modules

conversions
Type conversions for BLS12-381 multisignatures.

Structs

CombinedSignatureBytes
Wrapper for a serialized combined signature.

IndividualSignatureBytes
Wrapper for a serialized individual signature.

PopBytes
Wrapper for a serialized proof of possession.

PublicKeyBytes
Wrapper for a serialized public key.

SecretKeyBytes
Wrapper for a serialized secret key.

Type Definitions

CombinedPublicKey
A BLS combined public key is a curve point in the G2 group.

CombinedSignature
A BLS multisignature is a curve point in the G1 group.

IndividualSignature
A BLS signature is a curve point in the G1 group.

Pop
A BLS Proof of Possession is a curve point in the G1 group (a domain-separated signature on the public key).

PublicKey
A BLS public key is a curve point in the G2 group.

SecretKey
A BLS secret key is a field element.